👾 Updated Index, 404, and all their components
- Moved conway to shared/features
- Updated new, now it's index and old index is gone
- 404 Updated with Background Entertainment
- Color Scheme embedded in many components for Index
