+++ src/components/shared/features/conway/libs/conwayReducer.ts
