+++ src/components/shared/features/conway/libs/useConway.ts
