+++ src/components/shared/features/conway/libs/utils/constant.ts
