+++ src/components/shared/features/conway/libs/utils/types.ts
