+++ src/pages/index.tsx
@@ -1,31 +1,37 @@
 //
-//  home.tsx
-//  personal-blog
+//  new.tsx
+//  exclaimation
 //
-//  Created by d-exclaimation on 4:22 PM.
+//  Created by d-exclaimation on 1:26 PM.
 //  Copyright © 2021 d-exclaimation. All rights reserved.
 //
 
 import React from 'react';
-
-import {Grid, GridItem, Text, Box, VStack, Center} from '@chakra-ui/react';
-import RouteSideCar from '../components/shared/routes/RoutesSideBar';
-import Carousel from '../components/Carousel';
-import EpicProfile from '../components/EpicProfile';
-import MetaHead from '../components/shared/meta/MetaHead';
-import FooterDisclaimer from '../components/shared/meta/FooterDisclaimer';
-import LoadingScreen from '../components/shared/features/LoadingScreen';
-import GameOfLife from '../components/conway/GameOfLife';
-
-import {withUrqlClient} from 'next-urql';
+import {
+    Box,
+    Grid,
+    GridItem,
+    Text,
+    VStack,
+    Stack
+} from '@chakra-ui/react';
 import {createUrqlClient} from '../lib/server/createUrqlClient';
-import {useWindowSize} from '../lib/hooks/useWindow';
-import {useLanguagesQuery, useProfileQuery} from '../models/graphql/types';
+import {withUrqlClient} from 'next-urql';
 import {useRouter} from 'next/router';
+import {useLanguagesQuery, useProfileQuery} from '../models/graphql/types';
+import LoadingScreen from '../components/shared/features/LoadingScreen';
+import Hero from '../components/templates/Hero';
+import {allSections} from '../components/core/SectionChildren';
+import {darkMode} from '../constants/color.scheme';
+import BackgroundEntertainment from '../components/shared/features/BackgroundEntertainment';
+import {useResponsive} from '../lib/hooks/useResponsive';
+import ExBoxedIcon from '../components/shared/icons/ExBoxedIcon';
+import MetaHead from '../components/shared/meta/MetaHead';
 
+const DELAY = 50;
 
-const Index: React.FC = () => {
-    const window = useWindowSize();
+export const Index: React.FC = () => {
+    const {isPortrait} = useResponsive();
     const router = useRouter();
     const [{fetching, data, error}] = useProfileQuery();
     const [lang] = useLanguagesQuery();
@@ -47,19 +53,53 @@ const Index: React.FC = () => {
         return <LoadingScreen />;
     }
 
+    const sides = allSections(data.profile, lang.data.topLang, isPortrait);
 
     return (
-        <>
-            <MetaHead title={data.profile.name} description={'Welcome to the d-exclaimation developer website by vin aka d-exclaimation. This is the website / web app for all things related to me. My profiles, links, repos, projects, bios, and blogs, you named it it is probably here'}/>
-            <div className="App-header">
-                <GameOfLife />
-                <Grid
-                    gap={4}
+        <div className="New-header">
+            <MetaHead
+                title={data.profile.name}
+                description={
+                    'Welcome to the d-exclaimation developer website by vin aka d-exclaimation. This is the website / web app for all things related to me. My profiles, links, repos, projects, bios, and blogs, you named it it is probably here'
+                }
+            />
+            <BackgroundEntertainment/>
+            <Grid
+                h="96vh"
+                gap=".5rem"
+                templateAreas={`
+                    'a  b  c'
+                    'j  🌟 d'
+                    'i  🌟 e'
+                    'h  g  f'
+                `}
+                gridTemplateRows="repeat(4, 25%)"
+                gridTemplateColumns="20vmin auto 20vmin"
+            >
+                {sides.map(({val, children}, idx) => {
+                    const delay = (idx + 1) * DELAY;
+                    return (
+                        <GridItem
+                            className="New-Section"
+                            key={idx}
+                            gridArea={val}
+                            bg={darkMode}
+                            style={{animationDelay: `${delay}ms`}}
+                        >
+                            {children}
+                        </GridItem>
+                    );
+                })}
+                <GridItem
+                    className="New-Section"
+                    gridArea="🌟"
+                    style={{animationDelay: `${13 * DELAY}ms`}}
+                    bg={darkMode}
                 >
-                    <GridItem colSpan={5} rowSpan={2} mt={window.width < window.height ? '30vw' : '10vw'}>
-                        <VStack>
-                            <RouteSideCar/>
-                            <EpicProfile name={data.profile.name}/>
+                    <Stack direction={isPortrait ? 'column' : 'row-reverse'} align="center">
+                        <ExBoxedIcon />
+                        <VStack align={isPortrait ? 'center' : 'flex-end'}>
+                            <Hero title={data.profile.name} />
                             <Box m={2}>
                                 <Text
                                     align={'center'}
@@ -67,24 +107,10 @@ const Index: React.FC = () => {
                                 >{data.profile.bio}</Text>
                             </Box>
                         </VStack>
-                    </GridItem>
-                    <GridItem colSpan={1} ocacity={0} />
-                    <GridItem colSpan={3}>
-                        <Carousel
-                            github={data.profile}
-                            langName={lang.data.topLang.lang}
-                            percentage={lang.data.topLang.percentage}
-                        />
-                    </GridItem>
-                    <GridItem colSpan={1} ocacity={0} />
-                    <GridItem colSpan={5}>
-                        <Center>
-                            <FooterDisclaimer/>
-                        </Center>
-                    </GridItem>
-                </Grid>
-            </div>
-        </>
+                    </Stack>
+                </GridItem>
+            </Grid>
+        </div>
     );
 };
